Add files via upload
This is a parallelized version of the GUI-based version of OpenPIV in MATLAB. It has a speed increase roughly proportional to the number of cores on your machine, by using a parfor loop instead of a standard for loop. Some functionality that is lost in this version is that the visual representation of vectors in the GUI is not shown, and the sequential setting of 1-2, 2-3, 3-4 is the only option. The existing option for 1-2, 3-4, 5-6, is not available due to the way parfor works in MATLAB.